+++ README.md
@@ -18,7 +18,7 @@ The following table list all sample codes related to the spring boot integration
 |-------------------------------------------------------------------------------------------------------------------------|---------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------|-----------|
 | [Spring Batch Implementation](./batch-boot-jpa-sample)                                                                  | The application, demonstrates implementing Spring Batch 5 using simple config and creating batch tables using liquibase                                                                                                                                             | Completed |
 | [Archunit Implementation](./boot-api-archunit-sample)                                                                   | The application, demonstrates how to apply arch unit rules to the spring boot project                                                                                                                                                                               | Completed |
-| [Choas Engineering Principles](./boot-choas-monkey)                                                                     | The application, demonstrates how to apply choas engineering concepts to the spring boot project, test using Gatling to demonstrate the difference                                                                                                                  | WIP       |
+| [Chaos Engineering Principles](./boot-chaos-monkey)                                                                     | The application, demonstrates how to apply chaos engineering concepts to the spring boot project, test using Gatling to demonstrate the difference                                                                                                                  | WIP       |
 | [mongodb-elasticsearch-integration](./boot-mongodb-elasticsearch)                                                       | The application, demonstrates saving the data in MongoDb and then searching in ElasticSearch for quick Search, GeoSpatial Search                                                                                                                                    | WIP       |
 | [Rabbit Mq Implementation](./boot-rabbitmq-thymeleaf)                                                                   | The application, demonstrates how rabbitmq works with producer side acknowledgement                                                                                                                                                                                 | Completed |
 | [Rest API Documentation with examples](./boot-rest-docs-sample)                                                         | This application, demonstrates ability to generate pdf API documentation using spring rest docs                                                                                                                                                                     | Completed |
