+++ boot-chaos-monkey/src/main/java/com/example/choasmonkey/ChaosMonkeyApplication.java
@@ -1,15 +1,15 @@
-package com.example.graphql;
+package com.example.choasmonkey;
 
-import com.example.graphql.config.ApplicationProperties;
+import com.example.choasmonkey.config.ApplicationProperties;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.boot.context.properties.EnableConfigurationProperties;
 
 @SpringBootApplication
 @EnableConfigurationProperties({ApplicationProperties.class})
-public class Application {
+public class ChaosMonkeyApplication {
 
     public static void main(String[] args) {
-        SpringApplication.run(Application.class, args);
+        SpringApplication.run(ChaosMonkeyApplication.class, args);
     }
 }
