+++ boot-chaos-monkey/src/main/java/com/example/choasmonkey/config/logging/Loggable.java
