+++ boot-chaos-monkey/src/main/java/com/example/choasmonkey/utils/AppConstants.java
