+++ boot-chaos-monkey/src/test/java/com/example/choasmonkey/TestChaosMonkeyApplication.java
@@ -3,9 +3,9 @@ package com.example.choasmonkey;
 import com.example.choasmonkey.common.ContainerConfig;
 import org.springframework.boot.SpringApplication;
 
-public class TestChoasMonkeyApplication {
+public class TestChaosMonkeyApplication {
 
     public static void main(String[] args) {
-        SpringApplication.from(ChoasMonkeyApplication::main).with(ContainerConfig.class).run(args);
+        SpringApplication.from(ChaosMonkeyApplication::main).with(ContainerConfig.class).run(args);
     }
 }
