+++ graphql/boot-graphql-webflux/pom.xml
@@ -227,7 +227,7 @@
 				<configuration>
 					<java>
 						<palantirJavaFormat>
-							<version>2.47.0</version>
+							<version>2.50.0</version>
 						</palantirJavaFormat>
 						<importOrder />
 						<removeUnusedImports />
