+++ graphql/boot-graphql-webflux/src/main/java/com/example/graphql/GraphQLWebFluxApplication.java
@@ -1,15 +1,15 @@
-package com.example.choasmonkey;
+package com.example.graphql;
 
-import com.example.choasmonkey.config.ApplicationProperties;
+import com.example.graphql.config.ApplicationProperties;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.boot.context.properties.EnableConfigurationProperties;
 
 @SpringBootApplication
 @EnableConfigurationProperties({ApplicationProperties.class})
-public class ChoasMonkeyApplication {
+public class GraphQLWebFluxApplication {
 
     public static void main(String[] args) {
-        SpringApplication.run(ChoasMonkeyApplication.class, args);
+        SpringApplication.run(GraphQLWebFluxApplication.class, args);
     }
 }
