+++ graphql/boot-graphql-webflux/src/test/java/com/example/graphql/TestGraphQLWebFluxApplication.java
@@ -3,9 +3,11 @@ package com.example.graphql;
 import com.example.graphql.common.ContainerConfig;
 import org.springframework.boot.SpringApplication;
 
-public class TestApplication {
+public class TestGraphQLWebFluxApplication {
 
     public static void main(String[] args) {
-        SpringApplication.from(Application::main).with(ContainerConfig.class).run(args);
+        SpringApplication.from(GraphQLWebFluxApplication::main)
+                .with(ContainerConfig.class)
+                .run(args);
     }
 }
