+++ httpClients/boot-http-proxy/src/main/java/com/example/rest/proxy/HttpProxyApplication.java
@@ -1,15 +1,15 @@
-package com.example.keysetpagination;
+package com.example.rest.proxy;
 
-import com.example.keysetpagination.config.ApplicationProperties;
+import com.example.rest.proxy.config.ApplicationProperties;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.boot.context.properties.EnableConfigurationProperties;
 
 @SpringBootApplication
 @EnableConfigurationProperties({ApplicationProperties.class})
-public class Application {
+public class HttpProxyApplication {
 
     public static void main(String[] args) {
-        SpringApplication.run(Application.class, args);
+        SpringApplication.run(HttpProxyApplication.class, args);
     }
 }
