+++ httpClients/boot-http-proxy/src/test/java/com/example/rest/proxy/TestHttpProxyApplication.java
@@ -3,9 +3,9 @@ package com.example.rest.proxy;
 import com.example.rest.proxy.common.ContainersConfig;
 import org.springframework.boot.SpringApplication;
 
-class TestApplication {
+class TestHttpProxyApplication {
 
     public static void main(String[] args) {
-        SpringApplication.from(Application::main).with(ContainersConfig.class).run(args);
+        SpringApplication.from(HttpProxyApplication::main).with(ContainersConfig.class).run(args);
     }
 }
