+++ jpa/keyset-pagination/blaze-persistence/src/test/java/com/example/keysetpagination/TestBlazePersistenceApplication.java
@@ -3,10 +3,12 @@ package com.example.keysetpagination;
 import com.example.keysetpagination.common.ContainersConfig;
 import org.springframework.boot.SpringApplication;
 
-public class TestApplication {
+public class TestBlazePersistenceApplication {
 
     public static void main(String[] args) {
         System.setProperty("spring.profiles.active", "local");
-        SpringApplication.from(Application::main).with(ContainersConfig.class).run(args);
+        SpringApplication.from(BlazePersistenceApplication::main)
+                .with(ContainersConfig.class)
+                .run(args);
     }
 }
