+++ pom.xml
@@ -17,7 +17,7 @@
     <modules>
         <module>batch-boot-jpa-sample</module>
         <module>boot-api-archunit-sample</module>
-        <module>boot-choas-monkey</module>
+        <module>boot-chaos-monkey</module>
         <module>boot-grafana-lgtm</module>
         <module>boot-mongodb-elasticsearch</module>
         <module>boot-opensearch-sample</module>
