+++ r2dbc/boot-jooq-r2dbc-sample/pom.xml
@@ -332,7 +332,7 @@
                 <configuration>
                     <java>
                         <googleJavaFormat>
-                            <version>1.25.0</version>
+                            <version>1.25.2</version>
                             <style>AOSP</style>
                         </googleJavaFormat>
                     </java>
